requirements.txt
acme==0.31.0
asgiref==3.3.4
asn1crypto==0.24.0
blinker==1.4
certbot==0.31.0
certbot-apache==0.31.0
certifi==2018.8.24
chardet==3.0.4
cloud-init==20.2
ConfigArgParse==0.13.0
configobj==5.0.6
cryptography==2.6.1
Django==3.0
django-appconf==1.0.4
django-compressor==2.4.1
django-model-utils==4.1.1
fail2ban==0.10.2
future==0.16.0
httplib2==0.11.3
idna==2.6
Jinja2==2.10
josepy==1.1.0
jsonpatch==1.21
jsonpointer==1.10
jsonschema==2.6.0
MarkupSafe==1.1.0
mock==2.0.0
oauthlib==2.1.0
parsedatetime==2.4
pbr==4.2.0
Pillow==8.2.0
psycopg2==2.7
pycurl==7.43.0.2
PyJWT==1.7.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
PySimpleSOAP==1.16.2
python-apt==1.8.4.3
python-augeas==0.5.0
python-debian==0.1.35
python-debianbts==2.8.2
pytz==2019.1
PyYAML==3.13
rcssmin==1.0.6
reportbug===7.5.3-deb10u1
requests==2.21.0
requests-toolbelt==0.8.0
rjsmin==1.1.0
six==1.12.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.24.1
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
